Split Hedgebox into multiple files and fix groups
