+++ posthog/demo/products/hedgebox/models.py
@@ -3,97 +3,25 @@ import math
 from dataclasses import dataclass, field
 from decimal import Decimal
 from enum import Enum, auto
-from typing import Any, Dict, List, Optional, Set, Tuple, cast
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    Dict,
+    List,
+    Optional,
+    Set,
+    Tuple,
+    cast,
+)
 
 import pytz
 
-from posthog.constants import INSIGHT_TRENDS, TRENDS_LINEAR, TRENDS_WORLD_MAP
-from posthog.demo.matrix.matrix import Cluster, Matrix
 from posthog.demo.matrix.models import SimPerson, SimSessionIntent
-from posthog.demo.matrix.randomization import Industry
-from posthog.models import (
-    Action,
-    ActionStep,
-    Cohort,
-    Dashboard,
-    DashboardTile,
-    Experiment,
-    FeatureFlag,
-    Insight,
-    InsightViewed,
-)
 
-# This is a simulation of an online drive SaaS called Hedgebox
-# See this RFC for the reasoning behind it:
-# https://github.com/PostHog/product-internal/blob/main/requests-for-comments/2022-03-23-great-demo-data.md
+from .taxonomy import *
 
-# Simulation features:
-# - the product is used by lots of personal users, but businesses bring the most revenue
-# - most users are from the US, but there are blips all over the world
-# - timezones are accurate on the country level
-# - usage times are accurate taking into account time of day, timezone, and user profile (personal or business)
-# - Hedgebox is sponsoring the well-known YouTube channel about technology Marius Tech Tips - there's a landing page
-# - an experiment with a new signup page is running, and it's showing positive results
-# - Internet Explorer users do worse
-
-# See this flowchart for the layout of the product:
-# https://www.figma.com/file/nmvylkFx4JdTRDqyo5Vkb5/Hedgebox-Paths
-
-# URLs
-
-SITE_URL = "https://hedgebox.net"
-
-URL_HOME = f"{SITE_URL}/"
-URL_MARIUS_TECH_TIPS = f"{SITE_URL}/mariustechtips/"
-URL_PRICING = f"{SITE_URL}/pricing/"
-
-URL_SIGNUP = f"{SITE_URL}/signup/"
-URL_LOGIN = f"{SITE_URL}/login/"
-dyn_url_invite = lambda invite_id: f"{SITE_URL}/invite/{invite_id}/"
-
-URL_FILES = f"{SITE_URL}/files/"
-dyn_url_file = lambda file_id: f"{SITE_URL}/files/{file_id}/"
-
-URL_ACCOUNT_SETTINGS = f"{SITE_URL}/account/settings/"
-URL_ACCOUNT_BILLING = f"{SITE_URL}/account/billing/"
-URL_ACCOUNT_TEAM = f"{SITE_URL}/account/team/"
-
-# Event taxonomy
-
-EVENT_SIGNED_UP = "signed_up"  # Properties: from_invite
-EVENT_LOGGED_IN = "logged_in"  # No extra properties
-EVENT_LOGGED_OUT = "logged_out"  # No extra properties
-
-EVENT_UPLOADED_FILE = "uploaded_file"  # Properties: file_type, file_size_b
-EVENT_DOWNLOADED_FILE = "downloaded_file"  # Properties: file_type, file_size_b
-EVENT_DELETED_FILE = "deleted_file"  # Properties: file_type, file_size_b
-EVENT_SHARED_FILE_LINK = "shared_file_link"  # Properties: file_type, file_size_b
-
-EVENT_UPGRADED_PLAN = "upgraded_plan"  # Properties: previous_plan, new_plan
-EVENT_DOWNGRADED_PLAN = "downgraded_plan"  # Properties: previous_plan, new_plan
-
-EVENT_INVITED_TEAM_MEMBER = "invited_team_member"  # No extra properties
-EVENT_REMOVED_TEAM_MEMBER = "removed_team_member"  # No extra properties
-
-EVENT_PAID_BILL = "paid_bill"  # Properties: plan, amount_usd
-
-# Group taxonomy
-
-GROUP_TYPE_ACCOUNT = "account"  # Properties: name, industry, used_mb, plan, team_size
-
-# Feature flags
-
-FILE_PREVIEWS_FLAG_KEY = "file-previews"
-NEW_SIGNUP_PAGE_FLAG_KEY = "signup-page-4.0"
-NEW_SIGNUP_PAGE_FLAG_ROLLOUT_PERCENT = 50
-PROPERTY_NEW_SIGNUP_PAGE_FLAG = f"$feature/{NEW_SIGNUP_PAGE_FLAG_KEY}"
-SIGNUP_SUCCESS_RATE_TEST = 0.5794
-SIGNUP_SUCCESS_RATE_CONTROL = 0.4887
-
-# World properties
-
-# How many clusters should be companies (made up of business users) as opposed to social circles (personal users)
-COMPANY_CLUSTERS_PROPORTION = 0.2
+if TYPE_CHECKING:
+    from posthog.demo.products.hedgebox.matrix import HedgeboxCluster
 
 
 class HedgeboxSessionIntent(SimSessionIntent):
@@ -756,449 +684,3 @@ class HedgeboxPerson(SimPerson):
             for neighbor in cast(List[HedgeboxPerson], self.cluster._list_amenable_neighbors(self.x, self.y))
             if neighbor.is_invitable
         ]
-
-
-@dataclass
-class HedgdboxCompany:
-    name: str
-    industry: Industry
-
-
-class HedgeboxCluster(Cluster):
-    matrix: "HedgeboxMatrix"
-
-    MIN_RADIUS: int = 0
-    MAX_RADIUS: int = 6
-
-    # Properties
-    company: Optional[HedgdboxCompany]  # None means the cluster is a social circle instead of a company
-
-    # Internal state - plain
-    _business_account: Optional[HedgeboxAccount]  # In social circle clusters the person-level account is used
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        is_company = self.random.random() < COMPANY_CLUSTERS_PROPORTION
-        if is_company:
-            self.company = HedgdboxCompany(
-                name=self.finance_provider.company(), industry=self.properties_provider.industry()
-            )
-        else:
-            self.company = None
-        self._business_account = None
-
-    def __str__(self) -> str:
-        return self.company.name if self.company else f"Social Circle #{self.index+1}"
-
-    def radius_distribution(self) -> float:
-        return self.random.betavariate(1.5, 5)
-
-    def initation_distribution(self) -> float:
-        return self.random.betavariate(1.8, 1)
-
-
-class HedgeboxMatrix(Matrix):
-    PRODUCT_NAME = "Hedgebox"
-    CLUSTER_CLASS = HedgeboxCluster
-    PERSON_CLASS = HedgeboxPerson
-
-    new_signup_page_experiment_start: dt.datetime
-    new_signup_page_experiment_end: dt.datetime
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        # Start new signup page experiment roughly halfway through the simulation, end soon before `now`
-        self.new_signup_page_experiment_end = self.now - dt.timedelta(days=2, hours=3, seconds=43)
-        self.new_signup_page_experiment_start = self.start + (self.new_signup_page_experiment_end - self.start) / 2
-
-    def set_project_up(self, team, user):
-        super().set_project_up(team, user)
-
-        # Actions
-        interacted_with_file_action = Action.objects.create(
-            name="Interacted with file", team=team, description="Logged-in interaction with a file.", created_by=user,
-        )
-        ActionStep.objects.bulk_create(
-            (
-                ActionStep(action=interacted_with_file_action, event=EVENT_DELETED_FILE),
-                ActionStep(action=interacted_with_file_action, event=EVENT_UPLOADED_FILE),
-                ActionStep(action=interacted_with_file_action, event=EVENT_DOWNLOADED_FILE),
-                ActionStep(action=interacted_with_file_action, event=EVENT_SHARED_FILE_LINK),
-            )
-        )
-
-        # Cohorts
-        Cohort.objects.create(
-            team=team,
-            name="Signed-up users",
-            created_by=user,
-            groups=[{"properties": [{"key": "email", "type": "person", "value": "is_set", "operator": "is_set"}]}],
-        )
-        real_users_cohort = Cohort.objects.create(
-            team=team,
-            name="Real users",
-            description="People who don't belong to the Hedgebox team.",
-            created_by=user,
-            groups=[
-                {"properties": [{"key": "email", "type": "person", "value": "@hedgebox.net$", "operator": "not_regex"}]}
-            ],
-        )
-        team.test_account_filters = [{"key": "id", "type": "cohort", "value": real_users_cohort.pk}]
-
-        # Dashboard: Key metrics (project home)
-        key_metrics_dashboard = Dashboard.objects.create(
-            team=team, name="🔑 Key metrics", description="Company overview.", pinned=True
-        )
-        team.primary_dashboard = key_metrics_dashboard
-        weekly_signups_insight = Insight.objects.create(
-            team=team,
-            dashboard=key_metrics_dashboard,
-            saved=True,
-            name="Weekly signups",
-            filters={
-                "events": [{"id": EVENT_SIGNED_UP, "type": "events", "order": 0}],
-                "actions": [],
-                "display": TRENDS_LINEAR,
-                "insight": INSIGHT_TRENDS,
-                "interval": "week",
-                "date_from": "-1m",
-            },
-            last_modified_at=self.now - dt.timedelta(days=23),
-            last_modified_by=user,
-        )
-        DashboardTile.objects.create(
-            dashboard=key_metrics_dashboard,
-            insight=weekly_signups_insight,
-            color="blue",
-            layouts={
-                "sm": {"h": 5, "w": 6, "x": 0, "y": 0, "minH": 5, "minW": 3},
-                "xs": {"h": 5, "w": 1, "x": 0, "y": 0, "minH": 5, "minW": 3, "moved": False, "static": False},
-            },
-        )
-        signups_by_country_insight = Insight.objects.create(
-            team=team,
-            dashboard=key_metrics_dashboard,
-            saved=True,
-            name="Last month's signups by country",
-            filters={
-                "events": [{"id": EVENT_SIGNED_UP, "type": "events", "order": 0}],
-                "actions": [],
-                "display": TRENDS_WORLD_MAP,
-                "insight": INSIGHT_TRENDS,
-                "breakdown_type": "event",
-                "breakdown": "$geoip_country_code",
-                "date_from": "-1m",
-            },
-            last_modified_at=self.now - dt.timedelta(days=6),
-            last_modified_by=user,
-        )
-        DashboardTile.objects.create(
-            dashboard=key_metrics_dashboard,
-            insight=signups_by_country_insight,
-            layouts={
-                "sm": {"h": 5, "w": 6, "x": 6, "y": 0, "minH": 5, "minW": 3},
-                "xs": {"h": 5, "w": 1, "x": 0, "y": 5, "minH": 5, "minW": 3, "moved": False, "static": False},
-            },
-        )
-        activation_funnel = Insight.objects.create(
-            team=team,
-            dashboard=key_metrics_dashboard,
-            saved=True,
-            name="Activation",
-            filters={
-                "events": [
-                    {
-                        "custom_name": "Signed up",
-                        "id": EVENT_SIGNED_UP,
-                        "name": EVENT_SIGNED_UP,
-                        "type": "events",
-                        "order": 2,
-                    },
-                    {
-                        "custom_name": "Upgraded plan",
-                        "id": EVENT_UPGRADED_PLAN,
-                        "name": EVENT_UPGRADED_PLAN,
-                        "type": "events",
-                        "order": 4,
-                    },
-                ],
-                "actions": [
-                    {
-                        "id": interacted_with_file_action.pk,
-                        "name": interacted_with_file_action.name,
-                        "type": "actions",
-                        "order": 3,
-                    },
-                ],
-                "display": "FunnelViz",
-                "insight": "FUNNELS",
-                "interval": "day",
-                "funnel_viz_type": "steps",
-                "filter_test_accounts": True,
-                "date_from": "-1m",
-            },
-            last_modified_at=self.now - dt.timedelta(days=19),
-            last_modified_by=user,
-        )
-        DashboardTile.objects.create(
-            dashboard=key_metrics_dashboard,
-            insight=activation_funnel,
-            layouts={
-                "sm": {"h": 5, "w": 6, "x": 0, "y": 5, "minH": 5, "minW": 3},
-                "xs": {"h": 5, "w": 1, "x": 0, "y": 10, "minH": 5, "minW": 3, "moved": False, "static": False},
-            },
-        )
-        weekly_uploader_retention = Insight.objects.create(
-            team=team,
-            dashboard=key_metrics_dashboard,
-            saved=True,
-            name="App retention",
-            filters={
-                "period": "Week",
-                "display": "ActionsTable",
-                "insight": "RETENTION",
-                "properties": {
-                    "type": "AND",
-                    "values": [
-                        {
-                            "type": "AND",
-                            "values": [{"key": "email", "type": "person", "value": "is_set", "operator": "is_set"}],
-                        }
-                    ],
-                },
-                "target_entity": {
-                    "id": interacted_with_file_action.pk,
-                    "name": interacted_with_file_action.name,
-                    "type": "actions",
-                    "order": 0,
-                },
-                "retention_type": "retention_recurring",
-                "total_intervals": 9,
-                "returning_entity": {
-                    "id": interacted_with_file_action.pk,
-                    "name": interacted_with_file_action.name,
-                    "type": "actions",
-                    "order": 0,
-                },
-            },
-            last_modified_at=self.now - dt.timedelta(days=34),
-            last_modified_by=user,
-        )
-        DashboardTile.objects.create(
-            dashboard=key_metrics_dashboard,
-            insight=weekly_uploader_retention,
-            layouts={
-                "sm": {"h": 5, "w": 6, "x": 6, "y": 5, "minH": 5, "minW": 3},
-                "xs": {"h": 5, "w": 1, "x": 0, "y": 15, "minH": 5, "minW": 3, "moved": False, "static": False},
-            },
-        )
-
-        # Dashboard: Revenue
-        revenue_dashboard = Dashboard.objects.create(team=team, name="💸 Revenue", pinned=True)
-        monthly_app_revenue = Insight.objects.create(
-            team=team,
-            dashboard=revenue_dashboard,
-            saved=True,
-            name="Monthly app revenue",
-            filters={
-                "events": [
-                    {"id": EVENT_PAID_BILL, "type": "events", "order": 0, "math": "sum", "math_property": "amount_usd"}
-                ],
-                "actions": [],
-                "display": TRENDS_LINEAR,
-                "insight": INSIGHT_TRENDS,
-                "interval": "month",
-                "date_from": "-6m",
-            },
-            last_modified_at=self.now - dt.timedelta(days=29),
-            last_modified_by=user,
-        )
-        DashboardTile.objects.create(
-            dashboard=revenue_dashboard,
-            insight=monthly_app_revenue,
-            layouts={
-                "sm": {"h": 5, "w": 6, "x": 0, "y": 0, "minH": 5, "minW": 3},
-                "xs": {"h": 5, "w": 1, "x": 0, "y": 0, "minH": 5, "minW": 3, "moved": False, "static": False},
-            },
-        )
-
-        # Insight
-        Insight.objects.create(
-            team=team,
-            saved=True,
-            name="Homepage view to signup conversion",
-            filters={
-                "events": [
-                    {
-                        "custom_name": "Viewed homepage",
-                        "id": "$pageview",
-                        "name": "$pageview",
-                        "type": "events",
-                        "order": 0,
-                        "properties": [
-                            {"key": "$current_url", "type": "event", "value": URL_HOME, "operator": "exact",}
-                        ],
-                    },
-                    {
-                        "custom_name": "Viewed signup page",
-                        "id": "$pageview",
-                        "name": "$pageview",
-                        "type": "events",
-                        "order": 1,
-                        "properties": [
-                            {"key": "$current_url", "type": "event", "value": URL_SIGNUP, "operator": "regex",}
-                        ],
-                    },
-                    {"custom_name": "Signed up", "id": "signed_up", "name": "signed_up", "type": "events", "order": 2},
-                ],
-                "actions": [],
-                "display": "FunnelViz",
-                "insight": "FUNNELS",
-                "interval": "day",
-                "funnel_viz_type": "steps",
-                "filter_test_accounts": True,
-                "date_from": "-1m",
-            },
-            last_modified_at=self.now - dt.timedelta(days=19),
-            last_modified_by=user,
-        )
-        Insight.objects.create(
-            team=team,
-            saved=True,
-            name="User paths starting at homepage",
-            filters={
-                "date_to": None,
-                "insight": "PATHS",
-                "date_from": "-30d",
-                "edge_limit": 50,
-                "properties": {"type": "AND", "values": []},
-                "step_limit": 5,
-                "start_point": URL_HOME,
-                "funnel_filter": {},
-                "exclude_events": [],
-                "path_groupings": ["/files/*"],
-                "include_event_types": ["$pageview"],
-                "local_path_cleaning_filters": [],
-            },
-            last_modified_at=self.now - dt.timedelta(days=9),
-            last_modified_by=user,
-        )
-        Insight.objects.create(
-            team=team,
-            saved=True,
-            name="File interactions",
-            filters={
-                "events": [
-                    {"id": EVENT_UPLOADED_FILE, "type": "events", "order": 0},
-                    {"id": EVENT_DELETED_FILE, "type": "events", "order": 2},
-                    {"id": EVENT_DOWNLOADED_FILE, "type": "events", "order": 1},
-                ],
-                "actions": [],
-                "display": TRENDS_LINEAR,
-                "insight": INSIGHT_TRENDS,
-                "interval": "day",
-                "date_from": "-30d",
-            },
-            last_modified_at=self.now - dt.timedelta(days=13),
-            last_modified_by=user,
-        )
-
-        # InsightViewed
-        InsightViewed.objects.bulk_create(
-            (
-                InsightViewed(
-                    team=team,
-                    user=user,
-                    insight=insight,
-                    last_viewed_at=(
-                        self.now - dt.timedelta(days=self.random.randint(0, 3), minutes=self.random.randint(5, 60))
-                    ),
-                )
-                for insight in Insight.objects.filter(team=team)
-            )
-        )
-
-        # Feature flags
-        new_signup_page_flag = FeatureFlag.objects.create(
-            team=team,
-            key=FILE_PREVIEWS_FLAG_KEY,
-            name="File previews (ticket #2137). Work-in-progress, so only visible internally at the moment",
-            filters={
-                "groups": [
-                    {
-                        "properties": [
-                            {
-                                "key": "email",
-                                "type": "person",
-                                "value": [
-                                    "mark.s@hedgebox.net",
-                                    "helly.r@hedgebox.net",
-                                    "irving.b@hedgebox.net",
-                                    "dylan.g@hedgebox.net",
-                                ],
-                                "operator": "exact",
-                            }
-                        ]
-                    }
-                ]
-            },
-            created_by=user,
-            created_at=self.now - dt.timedelta(days=15),
-        )
-
-        # Experiments
-        new_signup_page_flag = FeatureFlag.objects.create(
-            team=team,
-            key=NEW_SIGNUP_PAGE_FLAG_KEY,
-            name="New sign-up flow",
-            filters={
-                "groups": [{"properties": [], "rollout_percentage": None}],
-                "multivariate": {
-                    "variants": [
-                        {"key": "control", "rollout_percentage": 100 - NEW_SIGNUP_PAGE_FLAG_ROLLOUT_PERCENT},
-                        {"key": "test", "rollout_percentage": NEW_SIGNUP_PAGE_FLAG_ROLLOUT_PERCENT},
-                    ]
-                },
-            },
-            created_by=user,
-            created_at=self.new_signup_page_experiment_start - dt.timedelta(hours=1),
-        )
-        Experiment.objects.create(
-            team=team,
-            name="New sign-up flow",
-            description="We've rebuilt our sign-up page to offer a more personalized experience. Let's see if this version performs better with potential users.",
-            feature_flag=new_signup_page_flag,
-            created_by=user,
-            filters={
-                "events": [
-                    {
-                        "id": "$pageview",
-                        "name": "$pageview",
-                        "type": "events",
-                        "order": 0,
-                        "properties": [
-                            {"key": "$current_url", "type": "event", "value": URL_SIGNUP, "operator": "exact",}
-                        ],
-                    },
-                    {"id": "signed_up", "name": "signed_up", "type": "events", "order": 1},
-                ],
-                "actions": [],
-                "display": "FunnelViz",
-                "insight": "FUNNELS",
-                "interval": "day",
-                "funnel_viz_type": "steps",
-                "filter_test_accounts": True,
-            },
-            parameters={
-                "feature_flag_variants": [
-                    {"key": "control", "rollout_percentage": 100 - NEW_SIGNUP_PAGE_FLAG_ROLLOUT_PERCENT},
-                    {"key": "test", "rollout_percentage": NEW_SIGNUP_PAGE_FLAG_ROLLOUT_PERCENT},
-                ],
-                "recommended_sample_size": int(len(self.clusters) * 0.43),
-                "recommended_running_time": None,
-                "minimum_detectable_effect": 1,
-            },
-            start_date=self.new_signup_page_experiment_start,
-            end_date=self.new_signup_page_experiment_end,
-            created_at=new_signup_page_flag.created_at,
-        )
